chore: CLRetryResult로 파일 분리 및 필요없는 코드 삭제
